+++ lithops/multiprocessing/popen_cloud.py
@@ -19,8 +19,6 @@ from lithops.wait import ALL_COMPLETED, ALWAYS
 
 from . import util
 
-__all__ = ['PopenCloud']
-
 
 #
 # Start child process using cloud
@@ -65,4 +63,4 @@ class PopenCloud(object):
 
     def _launch(self, process_obj):
         fn_args = [*process_obj._args, *process_obj._kwargs]
-        self.sentinel = self._executor.call_async(process_obj._target, fn_args)
+        self.sentinel = self._executor.call_async(process_obj._target, tuple(fn_args))