+++ lithops/serverless/backends/code_engine/code_engine.py
@@ -68,7 +68,7 @@ class CodeEngineBackend:
         try:
             self.region = self.cluster.split('//')[1].split('.')[1]
         except Exception:
-            self.region = ''
+            self.region = self.cluster.replace('http://', '').replace('https://', '')
 
         self.job_def_ids = set()
 
@@ -343,16 +343,16 @@ class CodeEngineBackend:
 
         return activation_id
 
-    def _create_job_definition(self, docker_image_name, runtime_memory, timeout):
+    def _create_job_definition(self, image_name, runtime_memory, timeout):
         """
         Creates a Job definition
         """
-        jobdef_name = self._format_jobdef_name(docker_image_name, runtime_memory)
+        jobdef_name = self._format_jobdef_name(image_name, runtime_memory)
 
         jobdef_res = yaml.safe_load(ce_config.JOBDEF_DEFAULT)
         jobdef_res['metadata']['name'] = jobdef_name
         container = jobdef_res['spec']['template']['containers'][0]
-        container['image'] = docker_image_name
+        container['image'] = '/'.join([self.code_engine_config['container_registry'], image_name])
         container['name'] = jobdef_name
         container['env'][0]['value'] = 'run'
         container['resources']['requests']['memory'] = '{}Mi'.format(runtime_memory)
@@ -379,7 +379,7 @@ class CodeEngineBackend:
             )
             logger.debug("response - {}".format(res))
         except Exception as e:
-            logger.debug(e)
+            raise e
 
         logger.debug('Job Definition {} created'.format(jobdef_name))
 
@@ -468,7 +468,7 @@ class CodeEngineBackend:
             try:
                 logger.debug("Retry attempt {} to read {}".format(retry, status_key))
                 json_str = internal_storage.get_data(key=status_key)
-                logger.debug("Found in attempt () to read {}".format(retry, status_key))
+                logger.debug("Found in attempt {} to read {}".format(retry, status_key))
                 runtime_meta = json.loads(json_str.decode("ascii"))
                 found = True
             except StorageNoSuchKeyError:
@@ -493,19 +493,25 @@ class CodeEngineBackend:
         self._delete_config_map(jobdef_name)
         return runtime_meta
 
-    def _generate_runtime_meta_service(self, docker_image_name, memory):
+    def _generate_runtime_meta_service(self, image_name, memory):
         """
-        Creates a service in CodeEngine based on the docker_image_name
+        Creates a service in CodeEngine based on the docker_image_name.
+
+        This is an alternative method to extract the runtime metadata.
+        Currently it is deactivated in favor of _generate_runtime_meta()
+        method which, for now, seems to be faster.
         """
-        logger.info("Extracting Python modules from: {}".format(docker_image_name))
+        logger.info("Extracting Python modules from: {}".format(image_name))
         svc_res = yaml.safe_load(kconfig.service_res)
 
-        service_name = docker_image_name.replace('/', '--').replace(':', '--')
+        service_name = image_name.replace('/', '--').replace(':', '--')
+        full_image_name = '/'.join([self.code_engine_config['container_registry'], image_name])
+
         svc_res['metadata']['name'] = service_name
         svc_res['metadata']['namespace'] = self.namespace
         svc_res['spec']['template']['spec']['timeoutSeconds'] = 30
         svc_res['spec']['template']['spec']['containerConcurrency'] = 1
-        svc_res['spec']['template']['spec']['containers'][0]['image'] = docker_image_name
+        svc_res['spec']['template']['spec']['containers'][0]['image'] = full_image_name
         svc_res['spec']['template']['spec']['containers'][0]['resources']['limits']['memory'] = '128Mi'
         svc_res['spec']['template']['spec']['containers'][0]['resources']['limits']['cpu'] = '0.1'
         svc_res['spec']['template']['spec']['containers'][0]['resources']['requests']['memory'] = '128Mi'
@@ -592,7 +598,7 @@ class CodeEngineBackend:
             self.coreV1Api.create_namespaced_config_map(namespace=self.namespace, body=cmap, field_manager=field_manager)
             logger.debug("ConfigMap {} for namespace {} created".format(config_name, self.namespace))
         except ApiException as e:
-            logger.warn("Exception when calling CoreV1Api->create_namespaced_config_map: %s\n" % e)
+            logger.warning("Exception when calling CoreV1Api->create_namespaced_config_map: %s\n" % e)
             if (e.status != 409):
                 raise Exception('Failed to create ConfigMap')
 
@@ -605,6 +611,6 @@ class CodeEngineBackend:
         try:
             logger.debug("Delete ConfigMap {} for namespace {}".format(config_name, self.namespace))
             api_response = self.coreV1Api.delete_namespaced_config_map(name=config_name, namespace=self.namespace, grace_period_seconds=grace_period_seconds)
-            logger.debug(" ConfigMap {} for namespace {} deleted with status {}".format(config_name, self.namespace, api_response.status))
+            logger.debug("ConfigMap {} for namespace {} deleted with status {}".format(config_name, self.namespace, api_response.status))
         except ApiException as e:
-            logger.warn("Exception when calling CoreV1Api->delete_namespaced_config_map: %s\n" % e)
+            logger.warning("Exception when calling CoreV1Api->delete_namespaced_config_map: %s\n" % e)