+++ lithops/standalone/backends/ibm_vpc/ibm_vpc.py
@@ -5,6 +5,11 @@ import time
 from lithops.constants import COMPUTE_CLI_MSG
 from lithops.util.ibm_token_manager import IBMTokenManager
 
+from ibm_vpc import VpcV1
+from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
+from ibm_cloud_sdk_core import ApiException
+import namegenerator
+
 logger = logging.getLogger(__name__)
 
 
@@ -19,6 +24,7 @@ class IBMVPCInstanceClient:
         self.region = self.endpoint.split('//')[1].split('.')[0]
         self.instance_id = self.config['instance_id']
         self.ip_address = self.config.get('ip_address', None)
+        self.vm_create_timeout = self.config.get('vm_create_timeout', 120)
 
         self.instance_data = None
 
@@ -29,6 +35,10 @@ class IBMVPCInstanceClient:
         self.session = requests.session()
 
         iam_api_key = self.config.get('iam_api_key')
+
+        authenticator = IAMAuthenticator(iam_api_key)
+        self.service = VpcV1('2020-06-02', authenticator=authenticator)
+
         token = self.config.get('token', None)
         token_expiry_time = self.config.get('token_expiry_time', None)
         api_key_type = 'IAM'
@@ -107,15 +117,123 @@ class IBMVPCInstanceClient:
             time.sleep(1)
             self.instance_data = self.get_instance()
 
+    def _generate_name(self, r_type):
+        return "lithops-" + namegenerator.gen() + "-" + r_type
+
+    def _create_instance(self):
+        # security_group_identity_model = {'id': 'r006-2d3cc459-bb8b-4ec6-a5fb-28e60c9f7d7b'}
+        security_group_identity_model = {'id': self.config['security_group_id']}
+
+        # subnet_identity_model = {'id': '0737-bbc80a8f-d46a-4cc6-8a5a-991daa5fc914'}
+        subnet_identity_model = {'id': self.config['subnet_id']}
+        
+        # key_identity_model = {'id': "r006-14719c2a-80cf-4043-8018-fa22d4ce1337"}
+        key_identity_model = {'id': self.config['key_id']}
+
+        volume_prototype_instance_by_image_context_model = {
+            'capacity': 100, 'iops': 10000, 'name': self._generate_name('volume'), 'profile': {'name': self.config['volume_tier_name']}}#''10iops-tier'}}
+
+        network_interface_prototype_model = {
+            'name': 'eth0', 'subnet': subnet_identity_model, 'security_groups': [security_group_identity_model]}
+        volume_attachment_prototype_instance_by_image = {
+            'delete_volume_on_instance_delete': True,
+            'name': self._generate_name('boot'),
+            'volume': volume_prototype_instance_by_image_context_model
+        }
+        instance_prototype_model = {
+            'keys': [key_identity_model], 'name': self._generate_name('instance')}
+        instance_prototype_model['profile'] = {'name': self.config['profile_name']}#"bx2-8x32"}
+
+        instance_prototype_model['resource_group'] = {'id': self.config['resource_group_id']}#"8145289ddf7047ea93fd2835de391f43"}
+        instance_prototype_model['vpc'] = {'id': self.config['vpc_id']}#"r006-afdd7b5d-059f-413f-a319-c0a38ef46824"}
+        instance_prototype_model['image'] = {'id': self.config['image_id']}#"r006-988caa8b-7786-49c9-aea6-9553af2b1969"}
+        instance_prototype_model['zone'] = {'name': self.config['zone_name']}#"us-south-3"}
+
+        instance_prototype_model['boot_volume_attachment'] = volume_attachment_prototype_instance_by_image
+        instance_prototype_model['primary_network_interface'] = network_interface_prototype_model
+
+        response = self.service.create_instance(instance_prototype_model)
+        return response.result
+        
+    def _create_and_attach_floating_ip(self, instance):
+        # allocate new floating ip
+        floating_ip_prototype_model = {}
+        floating_ip_prototype_model['name'] = self._generate_name('fip')
+        floating_ip_prototype_model['zone'] = {'name': self.config['zone_name']}#"us-south-3"}
+        floating_ip_prototype_model['resource_group'] = {'id': self.config['resource_group_id']}#"8145289ddf7047ea93fd2835de391f43"}
+
+        response = self.service.create_floating_ip(floating_ip_prototype_model)
+        floating_ip = response.result
+        
+        # attach floating ip
+        response = self.service.add_instance_network_interface_floating_ip(
+            instance['id'], instance['network_interfaces'][0]['id'], floating_ip['id'])
+
+        return floating_ip
+
+    def _wait_instance_running(self, instance_id):
+        """
+        Waits until the VM instance is running
+        """
+        logger.debug('Waiting VM instance to become running')
+
+        start = time.time()
+        while(time.time() - start < self.vm_create_timeout):
+            instance = self.service.get_instance(instance_id).result
+            if instance['status'] == 'running':
+                return True
+            time.sleep(1)
+
+        self.stop()
+        raise Exception('VM create failed, check logs and configurations')
+
+    def _delete_instance(self):
+        # delete floating ip
+        response = self.service.list_instance_network_interfaces(self.config['instance_id'])
+        for nic in response.result['network_interfaces']:
+            if 'floating_ips' in nic:
+                for fip in nic['floating_ips']:
+                    self.service.delete_floating_ip(fip['id'])
+                    print("floating ip {} been deleted".format(fip['address']))
+
+        # delete vm instance
+        self.service.delete_instance(self.config['instance_id'])
+        print("instance {} been deleted".format(self.config['instance_id']))
+
     def start(self):
         logger.info("Starting VM instance")
         self.create_instance_action('start')
         logger.debug("VM instance started successfully")
 
+    def create(self):
+        logger.info("Creating VM instance")
+        self.instance_id = None
+
+        try:
+            instance = self._create_instance()
+            floating_ip = self._create_and_attach_floating_ip(instance)['address']
+            logger.debug("VM {} created successfully with floating IP {}".format(instance['name'], floating_ip))
+
+            self.instance_id = instance['id']
+            self.config['instance_id'] = self.instance_id
+            self.config['ip_address'] = floating_ip
+
+            self._wait_instance_running(self.instance_id)
+            return self.instance_id, floating_ip
+        except Exception as e:
+            logger.error("There was an error trying to create the VM and  bind floating ip, deleting the vm {}".format(self.instance_id))
+            self._delete_instance()
+            raise e
+
     def stop(self):
-        logger.info("Stopping VM instance")
-        self.create_instance_action('stop')
-        logger.debug("VM instance stopped successfully")
+        if self.config['delete_on_dismantle']:
+            logger.info("Deleting VM instance")
+            self._delete_instance()
+            logger.debug("VM instance deleted successfully")
+        else:
+            logger.info("Stopping VM instance")
+            self.create_instance_action('stop')
+            logger.debug("VM instance stopped successfully")
 
     def get_runtime_key(self, runtime_name):
         runtime_key = os.path.join(self.name, self.ip_address,